v. 0.1.6.0 !! Multithreading is fully supported now (as far as I see, more research!) Ran cleaner in code for more readability. Added locks for thread-safe variable access.
Calculated support without major issues (on tested machine specs) is now about 45 jobs without critical issues.(though, a lot of IOExceptions due to a lot of thread access on the same file, this is ignorable due to the failsafe of writing to a string). This is about a 500% increase!

Supposedly, a dual core processor should support 50 jobs. A quadcore 100 jobs. Will need to do testing to see how this works.